backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
boto3==1.34.40
pydantic==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
